% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_ecs_folders.R
\name{config_ecs_folders}
\alias{config_ecs_folders}
\title{Configura las carpetas de proyecto de "Edad con Salud" (i.e., las carpetas
compartidas de OneDrive)}
\usage{
config_ecs_folders()
}
\value{
`TRUE` (invisible) si las carpetas se han configurado correctamente;
        `FALSE` (invisible) si hay algún problema durante la configuración
                de las carpetas.
}
\description{
Configura las carpetas de proyecto de "Edad con Salud" (i.e., las carpetas
compartidas de OneDrive)
}
\details{
Muestra dos ventanas de selección de ruta, una para cada carpeta
         ("Bases de Datos" y "Documentación", respectivamente).
         Por defecto intentará inicializar las ventanas en las rutas
         más habituales de "Edad con Salud", lo cual no debería dar problemas
         si las carpetas compartidas de OneDrive se han sincronizado en las
         rutas por defecto.
         Para instrucciones sobre cómo sincronizar las carpetas compartidas,
         ver el
         \href{https://bit.ly/33PJOJe}{Manual de sincronización de OneDrive}.
}
\examples{
\dontrun{config_ecs_folders()}
}
\seealso{
Other configuration functions: 
\code{\link{configure_ecs}()}
}
\concept{configuration functions}
