% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_ecs.R
\name{configure_ecs}
\alias{configure_ecs}
\title{Configura el entorno y comprueba la configuración del paquete `ecs.data`}
\usage{
configure_ecs(
  check_R_version = FALSE,
  check_CRAN_deps = FALSE,
  check_Statamarkdown = TRUE,
  check_Stata = TRUE,
  conf_folders = TRUE,
  confirmation = FALSE
)
}
\arguments{
\item{check_R_version}{(`FALSE`) valor lógico indicando si comprobar la
versión de R en uso.}

\item{check_CRAN_deps}{(`FALSE`) valor lógico indicando si comprobar e
instalar los paquetes del "Cran"}

\item{check_Statamarkdown}{(`TRUE`) valor lógico indicando si comprobar
e instalar el paquete "Statamarkdown"}

\item{check_Stata}{(`TRUE`) valor lógico indicando si comprobar la
instalación de Stata}

\item{conf_folders}{(`TRUE`) valor lógico indicando si configurar las
carpetas de proyecto de Edad con Salud}
}
\value{
`TRUE` (invisible) si el entorno se ha configurado correctamente;
        `FALSE` (invisible) si hay algún problema durante la configuración.
}
\description{
Configura el entorno y comprueba la configuración del paquete `ecs.data`
}
\details{
Hay cinco requisitos para que el paquete funcione correctamente;
  cada uno de ellos se comprobará si el argumento correspondiente es `TRUE`:
  - Se compureba que la versión de R es igual o superior a la 4.0.5; en caso
  contrario se emite un aviso, pero no se considera un problema de
  configuración.
  - Se comprueba si están instaladas las dependencias del 'CRAN';
  en caso de que falte alguna, se intenta instalar.
  Devuelve `FALSE` si falla la instalación de algún paquete.
  - Se comprueba si el paquete `Statamarkdown` está instalado;
  en caso contrario se intenta instalar. Devuelve `FALSE` si la instalación
  no se consigue completar.
  - Se comprueba que es posible encontrar una instalación de 'Stata'.
  En caso de que no, se emite un aviso explicando la configuración manual,
  y de devuelve `FALSE`.
  - Se intenta configurar las carpetas de proyecto de Edad con Salud
  (para hacerlo correctamente es necesario que las carpetas de OneDrive
  estén sincronizadas localmente). En caso de que al menos una de ellas no se
  pueda configurar, se devuelve `FALSE`.
}
\examples{
\dontrun{configure_ecs()}
}
\seealso{
Other configuration functions: 
\code{\link{config_ecs_folders}()}
}
\concept{configuration functions}
