% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_files.R
\name{get_projectfile_path}
\alias{get_projectfile_path}
\title{Devuelve la ruta de acceso a un archivo de proyecto, ya sea en el propio
proyecto, o en el paquete `ecs.data`}
\usage{
get_projectfile_path(
  file,
  where = c("project", "package"),
  proj_path = rstudioapi::getActiveProject()
)
}
\arguments{
\item{file}{Archivo al que se quiere acceder}

\item{where}{(`"project"`) Uno de "project", o "package"}

\item{proj_path}{(`rstudioapi::getActiveProject()`)
Directorio raíz del proyecto}
}
\value{
Ruta al archivo solicitado
}
\description{
Devuelve la ruta de acceso a un archivo de proyecto, ya sea en el propio
proyecto, o en el paquete `ecs.data`
}
