% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ecs_folder.R
\name{read_ecs_folder}
\alias{read_ecs_folder}
\title{Lee la carpeta de proyecto de "Edad con Salud" (i.e., carpeta
compartida de OneDrive) indicada y la devuelve como una ruta relativa al
"HOME" del usuario.}
\usage{
read_ecs_folder(
  folder = PROJECT_FOLDER_KEYS,
  run_config = FALSE,
  rel = TRUE,
  rel_stata = FALSE
)
}
\arguments{
\item{folder}{("DOC") cadena de caracteres indicando la
_clave_, es decir, si se desea obtener la ruta a la carpeta de datos
("DB") o de documentación ("DOC", por defecto).}

\item{run_config}{(`FALSE`) valor lógico indicando si ejecutar la
configuración de las carpetas, en caso de que el archivo de
configuración esté vacío. En caso de no ejecutarse la configuración
(por defecto) si el archivo de configuración está vacío termina con un
error.}

\item{rel}{(`TRUE`) valor lógico indicando si devolver la ruta relativa al
"HOME" de usuario. Windows suele considerar el "HOME" como una ruta
 del tipo `C:/Users/<username>/Documents`.}

\item{rel_stata}{(`FALSE`) valor lógico indicando si devolver la ruta
relativa al "HOME" de usuario entendido según Stata. Windows suele
considerar el "HOME" como una ruta del tipo
`C:/Users/<username>/Documents`, mientras que Stata considera el
"HOME" como `C:/Users/<username>`. No tiene ningún efecto si
`rel = FALSE`.}
}
\value{
ruta local a la carpeta del  proyecto, relativa al "HOME" del usuario
}
\description{
Lee la carpeta de proyecto de "Edad con Salud" (i.e., carpeta
compartida de OneDrive) indicada y la devuelve como una ruta relativa al
"HOME" del usuario.
}
\details{
Devuelve la ruta a la carpeta de OneDrive de Edad con Salud
         seleccionada. La carpeta (_bases de datos_ o _documentación_)
         se selecciona mediante la clave correspondiente en el parámetro
         `folder`.
         Para instrucciones sobre cómo sincronizar las carpetas compartidas,
         ver el
         \href{https://bit.ly/33PJOJe}{Manual de sincronización de OneDrive}.
}
