% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss_to_stata.R
\name{spss_to_stata}
\alias{spss_to_stata}
\title{Interfaz gráfico para transformar un archivo de datos de SPSS a formato Stata}
\usage{
spss_to_stata()
}
\description{
Interfaz gráfico para transformar un archivo de datos de SPSS a formato Stata
}
\details{
Puede dar algún fallo no controlado
         (no se ha probado con muchos casos de uso)
}
\examples{
ecs.data:::pss_to_stata()
}
