% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ecs_folder.R
\name{read_ecs_folder}
\alias{read_ecs_folder}
\title{Lee la carpeta de proyecto de "Edad con Salud" (i.e., carpeta
compartida de OneDrive) indicada}
\usage{
read_ecs_folder(folder = PROJECT_FOLDER_KEYS, run_config = FALSE)
}
\arguments{
\item{folder}{("DOC") cadena de caracteres indicando la
_clave_, es decir, si se desea obtener la ruta a la carpeta de datos
("DB") o de documentación ("DOC", por defecto).}

\item{run_config}{(`FALSE`) valor lógico indicando si ejecutar la
configuración de las carpetas, en caso de que el archivo de
configuración esté vacío. En caso de no ejecutarse la configuración
(por defecto) si el archivo de configuración está vacío termina con un
error.}
}
\value{
ruta local a la carpeta del  proyecto
}
\description{
Lee la carpeta de proyecto de "Edad con Salud" (i.e., carpeta
compartida de OneDrive) indicada
}
\details{
Devuelve la ruta a la carpeta de OneDrive de Edad con Salud
         seleccionada. La carpeta (_bases de datos_ o _documentación_)
         se selecciona mediante la clave correspondiente en el parámetro
         `folder`.
         Para instrucciones sobre cómo sincronizar las carpetas compartidas,
         ver el
         \href{https://bit.ly/33PJOJe}{Manual de sincronización de OneDrive}.
}
